fix: update all downstream code for widened struct fields (Task 5)
Bridge library:
- bridge_qemu.h: next_tag uint8_t->uint16_t, add tag_mask field,
  bridge_wait_completion tag param uint8_t->uint16_t
- bridge_qemu.c: init tag_mask=0x00FF, use tag_mask for tag wrapping
- bridge_vcs.c: cpl.tag cast to uint16_t, init requester_id/completer_id
  fields, msi_event_t.vector cast to uint16_t with requester_id=0
- trace_log.c: add requester_id/target_bdf to TLP format, requester_id/
  completer_id to CPL format, requester_id to DMA and MSI formats,
  update CSV header, fix dma timestamp format for uint32_t

Tests:
- Fix uint8_t->uint16_t tag casts in tcp_roundtrip and tcp_cross_machine
- Fix uint32_t->uint16_t vector casts in msi_roundtrip, qemu_integration,
  and tcp_cross_machine
- Fix dma_req_t.timestamp narrowing in qemu_integration

+++ tests/e2e/test_tcp_cross_machine.c
@@ -74,7 +74,7 @@ static int run_server(int port) {
         tlp_entry_t tlp;
         CHECK(t->recv_tlp(t, &tlp) == 0);
         CHECK(tlp.type == TLP_MRD);
-        CHECK(tlp.tag == (uint8_t)i);
+        CHECK(tlp.tag == (uint16_t)i);
 
         cpl_entry_t cpl;
         memset(&cpl, 0, sizeof(cpl));
@@ -105,7 +105,7 @@ static int run_server(int port) {
     for (int i = 0; i < 5; i++) {
         msi_event_t ev;
         CHECK(t->recv_msi(t, &ev) == 0);
-        CHECK(ev.vector == (uint32_t)(i + 1));
+        CHECK(ev.vector == (uint16_t)(i + 1));
     }
     printf("[Server] Test 4: PASS\n\n");
 
@@ -178,7 +178,7 @@ static int run_client(const char *host, int port) {
         tlp_entry_t tlp;
         memset(&tlp, 0, sizeof(tlp));
         tlp.type = TLP_MRD;
-        tlp.tag = (uint8_t)i;
+        tlp.tag = (uint16_t)i;
         tlp.len = 4;
         tlp.addr = (uint64_t)(0x1000 + i * 4);
         CHECK(t->send_tlp(t, &tlp) == 0);
@@ -187,7 +187,7 @@ static int run_client(const char *host, int port) {
         CHECK(t->recv_cpl(t, &cpl) == 0);
         uint64_t lat = now_us() - t0;
         lat_sum += lat;
-        CHECK(cpl.tag == (uint8_t)i);
+        CHECK(cpl.tag == (uint16_t)i);
         uint32_t val;
         memcpy(&val, cpl.data, 4);
         CHECK(val == 0xBEEF0000 + (uint32_t)i);
@@ -223,7 +223,7 @@ static int run_client(const char *host, int port) {
     /* Test 4: MSI */
     printf("[Client] Test 4: MSI event (5x)...\n");
     for (int i = 0; i < 5; i++) {
-        msi_event_t ev = { .vector = (uint32_t)(i + 1), .timestamp = now_us() };
+        msi_event_t ev = { .vector = (uint16_t)(i + 1), .timestamp = now_us() };
         CHECK(t->send_msi(t, &ev) == 0);
     }
     printf("[Client] Test 4: PASS\n\n");
